view can show but is not the same lockbook dir
