+++ clients/apple/ShareExtension-iOS/Info.plist
